Add option to keep development files

+++ Scripts/build.sh
@@ -24,7 +24,20 @@
 
 build
 
-try "Removing unneeded development files..." rm -Rfv "$DEPLOYROOT"/usr/include "$DEPLOYROOT"/usr/bin/suscan.status "$DEPLOYROOT"/usr/lib/pkgconfig
+ENVFILE="$DEPLOYROOT/env"
+
+function make_envfile() {
+    echo "PREFIX=\"$DEPLOYROOT\"" > "$ENVFILE"
+    echo "export PKG_CONFIG_PATH=\"\$PKG_CONFIG_PATH:\$PREFIX/usr/lib/pkgconfig\"" >> "$ENVFILE"
+    echo "alias qmake_sd=\"qmake SIGDIGGER_PREFIX=\\\"\$PREFIX\\\" SUWIDGETS_PREFIX=\\\"\$PREFIX\\\"\"" >> "$ENVFILE"
+}
+
+if [ "x$DEVEL" != "x" ]; then
+    try    "Creating environment file..." make_envfile
+    notice "  Environment file for plugin development: $ENVFILE"
+else
+    try "Removing unneeded development files..." rm -Rfv "$DEPLOYROOT"/usr/include "$DEPLOYROOT"/usr/bin/suscan.status "$DEPLOYROOT"/usr/lib/pkgconfig
+fi
 
 function make_startup_script() {
     # Create startup script
@@ -41,6 +54,7 @@ try "Creating startup script for suscli" make_startup_script suscli
 
 try "Moving files out of /usr..." mv "$DEPLOYROOT"/usr/* "$DEPLOYROOT"
 try "Remove empty /usr..." rmdir "$DEPLOYROOT"/usr
+try "Creating symlinks..." ln -s "$DEPLOYROOT" "$DEPLOYROOT/usr"
 try "Setting permissions to wrapper scripts..." chmod a+x "$DEPLOYROOT"/SigDigger "$DEPLOYROOT"/suscli
 echo
 echo "Done. SigDigger compiled succesfully in $DEPLOYROOT"
